+++ commands/read_test.go
@@ -346,6 +346,51 @@ func TestSelectImplicitFilter(t *testing.T) {
 	}
 }
 
+func TestSelectInteractive(t *testing.T) {
+	db, err := CreateDatabaseForTesting()
+	defer db.Close()
+	if err != nil {
+		t.Fatalf("CreateDatabaseForTesting() err = %q, want nil", err)
+	}
+	OldOpenDatabase := OpenDatabase
+	OpenDatabase = OpenDatabaseForTesting(db)
+	defer func() { OpenDatabase = OldOpenDatabase }()
+	OldCloseDatabase := CloseDatabase
+	CloseDatabase = CloseDatabaseForTesting
+	defer func() { CloseDatabase = OldCloseDatabase }()
+	err = initDatabase(db)
+	if err != nil {
+		t.Fatalf("initDatabase() = %q, want nil", err)
+	}
+	_, err = createPassword(db, "mymachine1", "myservice1", "myuser1", "mypassword1", "plain")
+	if err != nil {
+		t.Fatalf("createPassword() = %q, want nil", err)
+	}
+	_, err = createPassword(db, "mymachine2", "myservice2", "myuser2", "mypassword2", "plain")
+	if err != nil {
+		t.Fatalf("createPassword() = %q, want nil", err)
+	}
+	// Interactive search.
+	os.Args = []string{""}
+	inBuf := new(bytes.Buffer)
+	inBuf.Write([]byte("mymachine1" + "\n"))
+	outBuf := new(bytes.Buffer)
+
+	actualRet := Main(inBuf, outBuf)
+
+	// myservice1 is found, myservice2 is not found.
+	expectedRet := 0
+	if actualRet != expectedRet {
+		t.Fatalf("Main() = %q, want %q", actualRet, expectedRet)
+	}
+	expectedOutput := "Search term: "
+	expectedOutput += "machine: mymachine1, service: myservice1, user: myuser1, password type: plain, password: mypassword1\n"
+	actualOutput := outBuf.String()
+	if actualOutput != expectedOutput {
+		t.Fatalf("actualOutput = %q, want %q", actualOutput, expectedOutput)
+	}
+}
+
 func TestOpenCloseDatabase(t *testing.T) {
 	// Intentionally not mocking OpenDatabase and CloseDatabase in this test.
 	OldCommand := Command