+++ guide/src/news.md
@@ -3,6 +3,7 @@
 ## main
 
 - create: add interactive mode in case `--machine` or `--user` is not specified
+- search: add interactive mode when no search terms are specified
 
 ## 5.0